(1) separate transaction instance from home test complete
* be care that the log path is changed.
