+++ app/HouseholdLedger/urls.py
@@ -20,24 +20,26 @@ from rest_framework_jwt.views import refresh_jwt_token
 from rest_framework_swagger.views import get_swagger_view
 schema_view = get_swagger_view(url='http:/0.0.0.0:8000/')
 
-from home import views
+from home import views as home_views
+from transaction import views as transaction_views
 from rest_framework import routers
 
 router = routers.DefaultRouter()
-router.register(r'users', views.UserViewSet)
+router.register(r'users', home_views.UserViewSet)
 
 urlpatterns = [
-
+    url(r'^', include(router.urls)),
     path('admin/', admin.site.urls),
     url(r'^account/', include('rest_auth.urls'), name='account'),
-
     url(r'^account/registration/', include('rest_auth.registration.urls')),
-    url(r'^refresh-token/', refresh_jwt_token),
-    url(r'^home/$',views.home, name='home'),
-    url(r'^transaction/$',views.transaction, name='transaction'),
-    url(r'^boot/$',views.transaction_types, name='boot'),
     url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    url(r'^', include(router.urls)),
+    url(r'^refresh-token/', refresh_jwt_token),
+
+    url(r'^home/$',home_views.home, name='home'),
+    url(r'^boot/$',home_views.transaction_types, name='boot'),
+
+    url(r'^transaction/$',transaction_views.transaction, name='transaction'),
+
     url(r'^swagger/$', schema_view)
 
 ]
