+++ app/home/serializer.py
@@ -1,6 +1,5 @@
 from django.contrib.auth.models import User, Group
 from rest_framework import serializers
-from home.models import Transaction
 
 
 class UserSerializer(serializers.HyperlinkedModelSerializer):
@@ -10,16 +9,4 @@ class UserSerializer(serializers.HyperlinkedModelSerializer):
                   'username',
                   'email',
                   'groups'
-                  )
-
-class TransactionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Transaction
-        fields = ('user_id',
-                  'transaction_id',
-                  'transaction_name',
-                  'cost_amount',
-                  'transaction_type',
-                  'created_at',
-                  'updated_at',
                   )
